+++ tsconfig.json
@@ -31,8 +31,8 @@
       "~utils": ["components/utils/index"],
       "~utils/*": ["components/utils/*"],
       "~supported-documents": ["supported-documents/index"],
-      "~WorkflowEngine": ["WorkflowEngine/index"],
-      "~WorkflowEngine/*": ["WorkflowEngine/*"],
+      "~workflow-engine": ["workflow-engine/index"],
+      "~workflow-engine/*": ["workflow-engine/*"],
       "~auth-sdk/*": ["../auth-sdk/FaceTec/*"]
     },
     "resolveJsonModule": true,
